feat: abstract filter logic and deduplicate
